# Add BANIDX facility to speed up BAN loading